Loot table fix
