+++ src/main/java/net/jesteur/me/block/special/toggeable_lights/AbstractToggeableLightBlock.java
@@ -1,30 +1,137 @@
 package net.jesteur.me.block.special.toggeable_lights;
 
 import net.minecraft.block.*;
+import net.minecraft.entity.ai.pathing.NavigationType;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.projectile.ProjectileEntity;
+import net.minecraft.fluid.FluidState;
+import net.minecraft.fluid.Fluids;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.Items;
+import net.minecraft.nbt.NbtCompound;
 import net.minecraft.particle.ParticleTypes;
+import net.minecraft.registry.tag.ItemTags;
+import net.minecraft.resource.featuretoggle.FeatureSet;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
+import net.minecraft.state.StateManager;
 import net.minecraft.state.property.BooleanProperty;
+import net.minecraft.state.property.IntProperty;
 import net.minecraft.state.property.Properties;
+import net.minecraft.state.property.Property;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.Hand;
 import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Direction;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.util.math.random.Random;
-import net.minecraft.world.World;
-import net.minecraft.world.WorldAccess;
+import net.minecraft.world.*;
 import net.minecraft.world.event.GameEvent;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Collections;
 
 public abstract class AbstractToggeableLightBlock extends Block {
-
+    public static final BooleanProperty LIT;
+    public static final IntProperty LEVEL_15;
+    public static final BooleanProperty WATERLOGGED;
     protected AbstractToggeableLightBlock(Settings settings) {
         super(settings);
     }
 
+    @Override
+    protected void appendProperties(StateManager.Builder<Block, BlockState> builder) {
+        builder.add(new Property[]{LEVEL_15, LIT, WATERLOGGED});
+    }
+    protected Iterable<Vec3d> getParticleOffsets(BlockState state) {
+        return null;
+    }
 
+    @Override
+    public BlockState getAppearance(BlockState state, BlockRenderView renderView, BlockPos pos, Direction side, @Nullable BlockState sourceState, @Nullable BlockPos sourcePos) {
+        return super.getAppearance(state, renderView, pos, side, sourceState, sourcePos);
+    }
+
+    public ItemStack getPickStack(BlockView world, BlockPos pos, BlockState state) {
+        return addNbtForLevel(super.getPickStack(world, pos, state), (Integer)state.get(LEVEL_15));
+    }
+
+    public static ItemStack addNbtForLevel(ItemStack stack, int level) {
+        if (level != 15) {
+            NbtCompound nbtCompound = new NbtCompound();
+            nbtCompound.putString(LEVEL_15.getName(), String.valueOf(level));
+            stack.setSubNbt("BlockStateTag", nbtCompound);
+        }
+        return stack;
+    }
+
+    @Override
+    public boolean isEnabled(FeatureSet enabledFeatures) {
+        return super.isEnabled(enabledFeatures);
+    }
+
+    @Override
+    public boolean canPlaceAt(BlockState state, WorldView world, BlockPos pos) {
+        Direction direction = attachedDirection(state).getOpposite();
+        return Block.sideCoversSmallSquare(world, pos.offset(direction), direction.getOpposite());
+    }
+
+    @Override
+    public BlockState getStateForNeighborUpdate(BlockState state, Direction direction, BlockState neighborState, WorldAccess world, BlockPos pos, BlockPos neighborPos) {
+        if (state.get(WATERLOGGED)) {
+            world.scheduleFluidTick(pos, Fluids.WATER, Fluids.WATER.getTickRate(world));
+            extinguish(null, state, world, pos);
+        }
+
+        return attachedDirection(state).getOpposite() == direction && !state.canPlaceAt(world, pos) ? Blocks.AIR.getDefaultState() : super.getStateForNeighborUpdate(state, direction, neighborState, world, pos, neighborPos);
+    }
+
+    protected abstract Direction attachedDirection(BlockState state);
+
+    @Override
+    public FluidState getFluidState(BlockState state) {
+        return state.get(WATERLOGGED) ? Fluids.WATER.getStill(false) : super.getFluidState(state);
+    }
+    @Override
+    public boolean canPathfindThrough(BlockState state, BlockView world, BlockPos pos, NavigationType type) { return false; }
+
+    @Override
+    public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
+        if (!world.isClient && player.getAbilities().allowModifyWorld) {
+            ItemStack itemStack = player.getStackInHand(hand);
+            if(state.get(LIT) && itemStack.isIn(ItemTags.SHOVELS)){
+                extinguish(null, state, world, pos);
+            } else if(!state.get(LIT) && itemStack.isOf(Items.FLINT_AND_STEEL) ||itemStack.isOf(Items.TORCH)){
+                setLit(world, state, pos, true);
+            }
+        }
+        return ActionResult.SUCCESS;
+    }
+
+    public static boolean emitsLight(BlockState state){
+        return state.get(LIT);
+    }
+
+    protected static void setLit(WorldAccess world, BlockState state, BlockPos pos, boolean lit) {
+        world.setBlockState(pos, state.with(LIT, lit).cycle(LEVEL_15), 2 | 3);
+        if(lit){
+            world.playSound((PlayerEntity)null, pos, SoundEvents.ITEM_FLINTANDSTEEL_USE, SoundCategory.BLOCKS, 1.5F, 1.0F);
+        }
+    }
+
+    protected static void extinguish(@Nullable PlayerEntity player, BlockState state, WorldAccess world, BlockPos pos) {
+        setLit(world, state, pos, false);
+
+        world.playSound((PlayerEntity)null, pos, SoundEvents.BLOCK_CANDLE_EXTINGUISH, SoundCategory.BLOCKS, 1.5F, 1.0F);
+        world.emitGameEvent(player, GameEvent.BLOCK_CHANGE, pos);
+    }
+
+    static {
+        LIT = Properties.LIT;
+        LEVEL_15 = Properties.LEVEL_15;
+        WATERLOGGED = Properties.WATERLOGGED;
+    }
 
 }
