+++ src/main/resources/data/me/loot_tables/blocks/skull_lantern_wall.json
@@ -1,0 +1,14 @@
+{
+  "type": "minecraft:block",
+  "pools": [
+    {
+      "rolls": 1,
+      "entries": [
+        {
+          "type": "minecraft:item",
+          "name": "me:skull_lantern"
+        }
+      ]
+    }
+  ]
+}
